requirements.txt
streamlit==1.38.0
streamlit-drawable-canvas==0.9.3
PyMuPDF==1.24.10
pillow==10.4.0
numpy==2.0.2
